@@ -7,22 +7,24 @@
     "test": "tests"
   },
   "scripts": {
-    "start": "node ./server/server.js",
+    "start": "node ./server_built/server/server.js",
     "build": "node ./node_modules/webpack/bin/webpack.js -w",
     "build-prod": "node ./node_modules/webpack/bin/webpack.js",
-    "seed": "node ./server/seed.js seed",
-    "seedDebug": "node %NODE_DEBUG_OPTION% ./server/seed.js seed",
+    "seed": "node ./server_built/server/seed.js seed",
     "postinstall": "patch-package"
   },
   "author": "",
   "license": "ISC",
   "devDependencies": {
+    "@types/js-cookie": "^3.0.2",
     "babel-cli": "^6.26.0",
     "css-loader": "^0.28.11",
     "style-loader": "^0.19.1",
+    "ts-loader": "^9.3.1",
     "url-loader": "^1.0.1",
-    "webpack": "^3.12.0",
-    "webpack-bundle-analyzer": "^3.3.2"
+    "webpack": "^5.74.0",
+    "webpack-bundle-analyzer": "^3.3.2",
+    "webpack-cli": "^4.10.0"
   },
   "dependencies": {
     "@babel/core": "^7.0.0-beta.51",
@@ -138,6 +140,7 @@
     "sqlite3": "^4.2.0",
     "storage-factory": "^0.1.0",
     "tether": "^1.4.4",
+    "typescript": "^4.7.4",
     "uuid": "^3.3.0",
     "whatwg-fetch": "^2.0.4",
     "winston": "^3.0.0-rc5",
