@@ -1,6 +1,6 @@
 const google = require('googleapis');
 const gmail = google.gmail('v1');
-const config = require('../config.js');
+const { config } = require('../config');
 const log = require('./logging.js');
 const db = require('./orm.js');
 const Op = db.Sequelize.Op;
