@@ -1,4 +1,4 @@
-const config = require('../config');
+const { config } = require('../config');
 const passport = require('passport');
 const GoogleStrategy = require('passport-google-oauth20').Strategy;
 const LocalStrategy = require('passport-local').Strategy;
